restructur: next pages폴더로 ssr 로직 이동
