+++ src/pages/app/plans/index.tsx
@@ -1,23 +1,18 @@
 import React, { useState, useEffect } from 'react';
-import type { NextPage, GetServerSideProps } from 'next';
-//import axios from 'axios';
-//import instance from '@/utils/axios';
-import { ssrInstance } from '@/shared/utils/axiosSsr';
+
 import { SortOption } from '@/shared/types/reviewType';
 import { useCursorInfiniteScroll } from '@/shared/hooks/useCursorInfiniteScrollPlans';
-import { PlanDataWithCategory, PlanListResponse } from '@/shared/types/plans';
+import { PlanDataWithCategory } from '@/shared/types/plans';
 import { RegionOption, SubRegionOption } from '@/shared/types/reviewType';
 import Tabs from '@/entities/plan/plans/tab/Tabs';
 import RenderTabContent from '@/entities/plan/plans/RenderTabContent';
-
-//const baseUrl = process.env.NEXT_PUBLIC_BASE_URL;
 
 interface HomeProps {
   initialPlans: PlanDataWithCategory[];
   initialCursor: number | null;
 }
 
-export const Home: NextPage<HomeProps> = ({ initialPlans, initialCursor }) => {
+export const Home = ({ initialPlans, initialCursor }: HomeProps) => {
   //상태관리
   const [plans, setPlans] = useState<PlanDataWithCategory[]>(initialPlans);
   const [cursor, setCursor] = useState<number | null | undefined>(
@@ -117,42 +112,4 @@ export const Home: NextPage<HomeProps> = ({ initialPlans, initialCursor }) => {
       <div ref={loaderRef} className="h-12"></div>
     </div>
   );
-};
-
-export const getServerSideProps: GetServerSideProps = async ({ req }) => {
-  try {
-    const cookie = req.headers.cookie || ''; // 요청에서 쿠키 가져오기
-    const isLoggedIn = !!cookie.includes('accessToken'); // accessToken 존재 여부로 로그인 상태 판별
-
-    //console.log('SSR 초기 데이터 요청 실행');
-    //console.log('SSR 요청 쿠키:', cookie || '없음');
-
-    const res = await ssrInstance(cookie).get<PlanListResponse>(
-      `/api/plans?size=10&sort=default`,
-      {
-        headers: isLoggedIn ? { Cookie: cookie } : {}, // 로그인 시 쿠키 포함
-        withCredentials: isLoggedIn, // 로그인 여부에 따라 withCredentials 설정
-      },
-    );
-
-    //onsole.log('SSR API 응답 데이터:', res.data);
-    const data = res.data;
-    const initialPlans: PlanDataWithCategory[] = data.data.planList.map(
-      (item) => ({ ...item }),
-    );
-    const nextCursor = data.data.nextCursor;
-    return {
-      props: {
-        initialPlans,
-        initialCursor: nextCursor !== undefined ? nextCursor : null,
-      },
-    };
-  } catch (error) {
-    console.error('초기 데이터 로딩 실패:', error);
-    return {
-      props: {
-        initialPlans: [],
-      },
-    };
-  }
 };
